[#92] refactor: ListPage, PapersPage 컴포넌트 분리
