+++ src/pages/ListPage.jsx
@@ -5,8 +5,11 @@ import useDeviceType from 'hooks/useDeviceType';
 import { Link } from 'react-router-dom';
 import routes from 'utils/constants/routes';
 import GNB from 'components/commons/header/GNB';
-import Button from 'components/commons/buttons/Button';
 import { useGetRecipientsInOrderQuery } from 'hooks/queries/useListQuery';
+import Footer from 'components/list/Footer';
+import ListPageMainContainer from 'components/list/ListPageMainContainer';
+import ListName from 'components/list/ListName';
+import PageChangeButton from 'components/list/PageChangeButton';
 
 const ListPage = () => {
   const deviceType = useDeviceType();
@@ -26,14 +29,14 @@ const ListPage = () => {
   return (
     <div>
       <GNB />
-      <MainContainer>
+      <ListPageMainContainer>
         <SectionConainer>
-          <StyledHtag>
+          <ListName>
             인기 롤링 페이퍼 🔥
             <Link to={`${routes.papers}/like`}>
-              <ChangeLayoutButton>전체 보기</ChangeLayoutButton>
+              <PageChangeButton>전체 보기</PageChangeButton>
             </Link>
-          </StyledHtag>
+          </ListName>
           <CardList
             carouselMargin={
               deviceType === 'PC' ? 0 : deviceType === 'Tablet' ? 2.4 : 2
@@ -44,12 +47,12 @@ const ListPage = () => {
           />
         </SectionConainer>
         <SectionConainer>
-          <StyledHtag>
+          <ListName>
             최근에 만든 롤링 페이퍼 ⭐️
             <Link to={`${routes.papers}/date`}>
-              <ChangeLayoutButton>전체 보기</ChangeLayoutButton>
+              <PageChangeButton>전체 보기</PageChangeButton>
             </Link>
-          </StyledHtag>
+          </ListName>
           <CardList
             carouselMargin={
               deviceType === 'PC' ? 0 : deviceType === 'Tablet' ? 2.4 : 2
@@ -59,115 +62,16 @@ const ListPage = () => {
             isError={isLastestListError}
           />
         </SectionConainer>
-      </MainContainer>
-      <StyledFooter>
-        <Link to={routes.post}>
-          <ListPageButton>나도 만들어보기</ListPageButton>
-        </Link>
-      </StyledFooter>
+      </ListPageMainContainer>
+      <Footer />
     </div>
   );
 };
 
 export default ListPage;
 
-export const MainContainer = styled.main`
-  display: flex;
-  flex-direction: column;
-  gap: 5.4rem;
-  padding-top: 4rem;
-  padding-bottom: 14.6rem;
-
-  @media (min-width: 48rem) {
-    gap: 3rem;
-    padding-top: 5rem;
-    padding-bottom: 23.6rem;
-  }
-
-  @media (min-width: 75rem) {
-    padding-bottom: 14.4rem;
-  }
-`;
-
 const SectionConainer = styled.section`
   @media (min-width: 75rem) {
     margin: 0 auto;
-  }
-`;
-
-export const StyledHtag = styled.h1`
-  position: relative;
-  color: ${({ theme }) => theme.black};
-  font-size: 2rem;
-  font-weight: 600;
-  line-height: 3rem;
-  margin-bottom: 0.2rem;
-  margin-left: 2rem;
-
-  @media (min-width: 48rem) {
-    font-size: 2.4rem;
-    font-weight: 700;
-    line-height: 3.6rem;
-    letter-spacing: -0.024rem;
-    margin-bottom: 0.6rem;
-    margin-left: 2.4rem;
-  }
-
-  @media (min-width: 75rem) {
-    margin-left: 0;
-  }
-`;
-
-export const StyledFooter = styled.footer`
-  position: fixed;
-  width: 100%;
-  bottom: 0;
-  padding: 2.4rem 2rem;
-
-  @media (min-width: 48rem) {
-    padding: 2.4rem;
-  }
-
-  @media (min-width: 75rem) {
-    padding-right: 0;
-    padding-left: 0;
-  }
-`;
-
-export const ListPageButton = styled(Button)`
-  position: static;
-  width: 100%;
-  height: 5.6rem;
-  font-size: 1.8rem;
-  font-weight: 400;
-  letter-spacing: -0.018rem;
-  line-height: 2.8rem;
-
-  @media (min-width: 48rem) {
-    font-weight: 700;
-  }
-
-  @media (min-width: 75rem) {
-    width: 28rem;
-    display: block;
-    margin: 0 auto;
-  }
-`;
-
-export const ChangeLayoutButton = styled(Button)`
-  position: absolute;
-  top: 0;
-  right: 1.2rem;
-  width: auto;
-  height: auto;
-  padding: 0.8rem;
-
-  @media (min-width: 48rem) {
-    right: 2.4rem;
-    padding: 1rem;
-  }
-
-  @media (min-width: 75rem) {
-    right: 1rem;
   }
 `;
